tests/texturing/getteximage-targets: fix int-to-pointer cast
tests/texturing/getteximage-targets.c:198:39: warning: cast to pointer
from integer of different size [-Wint-to-pointer-cast]

Use uintptr_t which is guaranteed to be the same size as a pointer.

Fixes: b13c204093c3 ("getteximage-targets: Add read texture image into PBO case for each target")
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/1127>

+++ tests/texturing/getteximage-targets.c
@@ -195,7 +195,7 @@ getTexImage(bool doPBO, GLenum target, GLubyte data[][IMAGE_SIZE],
 	for (i = 0; i < num_faces; i++) {
 		if (doPBO) {
 			glGetTexImage(target + i, 0, GL_RGBA, GL_UNSIGNED_BYTE,
-				      (GLvoid *) (long) (i * layer_size));
+				      (GLvoid *) (uintptr_t) (i * layer_size));
 		} else {
 			glGetTexImage(target + i, 0, GL_RGBA, GL_UNSIGNED_BYTE,
 				      data2[i]);
